requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
huggingface-hub
python-dotenv
openpyxl